fix: ErrorCode의 code 규칙 변경
